Add severity info to logstash and filter out DEBUG lines
This adds severity as a logstash field for every oslo formatted
log line, and does not add any lines which are at DEBUG level.

This means we no longer rely on the level=INFO query paremeter
in order to remove DEBUG lines, so we will avoid sending them
to logstash regardless of whether os-loganalyze is used.

Change-Id: I8c4ac76a7fa0c3badd82fc7c54959ef6eb052732