Install statsd version 2.1.2
The logstash log processing and subunit2sql tooling is often paired
together on servers. Currently subunit2sql depends on
os-performance-tools which depend on statsd<3.0. That means our package
resource here for statsd that tried to install latest conflicts with our
usage of subunit2sql on the same server.

Avoid this conflict by installing 2.1.2 for geard and then subunit2sql
will be happy too.

Change-Id: I3ac04cb93025ae2e2115ed23ba4927c2060f6dc8